+++ lib/src/Main/dialCodeSelectorView.dart
@@ -1,7 +1,6 @@
 import 'package:dialcodeselector/src/Model/country.dart';
 import 'package:dialcodeselector/src/Model/dialCodeSelectorTheme.dart';
 import 'package:dialcodeselector/src/ViewModel/countryViewModel.dart';
-import 'package:dialcodeselector/src/Views/countriesMainView.dart';
 import 'package:flutter/material.dart';
 
 class DialCodeSelector {
@@ -10,21 +9,11 @@ class DialCodeSelector {
   static Future<void> selectCountry(BuildContext context,
       {String? initialShortName,
       DialCodeSelectorTheme? selectorTheme,
-      Function(Country selectedCountry)? onCountrySelected}) async {
-    Country? country = await showModalBottomSheet(
-        isScrollControlled: true,
-        backgroundColor: Colors.transparent,
-        barrierColor: Colors.black.withOpacity(0.75),
-        shape: const RoundedRectangleBorder(
-            borderRadius: BorderRadius.vertical(top: Radius.circular(7))),
-        context: context,
-        builder: (context) {
-          return CountriesMainView(
-              dialCodeSelectorTheme: selectorTheme,
-              initialShortName: initialShortName);
-        });
-    if (country == null) return;
-    onCountrySelected?.call(Country.forCallBack(country));
+      Function(Country? selectedCountry)? onCountrySelected}) async {
+    _viewModel.selectCountry(context,
+        initialShortName: initialShortName,
+        selectorTheme: selectorTheme,
+        onCountrySelected: onCountrySelected);
   }
 
   static Future<void> init() async => _viewModel.getCountries;