+++ lib/src/Model/country.dart
@@ -15,13 +15,20 @@ class Country {
         countryName: json['name']);
   }
 
-  /// use this factory to use asset images in this package, by passing Country
-  factory Country.forCallBack(Country? country) {
+//USED WHEN COUNTRY IS NOT FOUND BY EITHER DIALCODE OR SHORTNAME
+  static Country get _defaultCountry => Country(
+      countryName: "Ghana",
+      icon: "packages/${PackageUtils.packageName}/asset/images/gh.png",
+      countryShortName: "GH",
+      dialCode: "+233");
+
+  static Country? forCallBack(Country? country) {
+    if (country == null) return _defaultCountry;
     return Country(
-        countryShortName: country?.countryShortName,
-        dialCode: country?.dialCode,
+        countryShortName: country.countryShortName,
+        dialCode: country.dialCode,
         icon:
-            "packages/${PackageUtils.packageName}/asset/images/${country?.countryShortName?.toLowerCase()}.png",
-        countryName: country?.countryName);
+            "packages/${PackageUtils.packageName}/asset/images/${country.countryShortName?.toLowerCase()}.png",
+        countryName: country.countryName);
   }
 }