+++ lib/src/Model/dialCodeSelectorTheme.dart
@@ -14,9 +14,11 @@ class DialCodeSelectorTheme {
   final Color? searchFieldBorderColor;
   final Color? searchFieldTextColor;
   final Color? checkedBoxColor;
+  final double? sheetCornerRadius;
   DialCodeSelectorTheme(
       {this.showCountriesOnly = false,
       this.title = "Select Country",
+      this.sheetCornerRadius = 11,
       this.titleFontWeight = FontWeight.bold,
       this.searchFieldBorderColor = PackageUtils.textColorBlack,
       this.searchFieldTextColor = PackageUtils.textColorBlack,