+++ lib/src/ViewModel/countryViewModel.dart
@@ -1,6 +1,8 @@
 import 'package:dialcodeselector/src/Model/country.dart';
 import 'package:dialcodeselector/src/Model/dialCodeSelectorTheme.dart';
 import 'package:dialcodeselector/src/Service/dialCodeSelectorService.dart';
+import 'package:dialcodeselector/src/Views/countriesMainView.dart';
+import 'package:flutter/material.dart';
 
 class CountryViewModel {
   CountryViewModel._();
@@ -22,6 +24,27 @@ class CountryViewModel {
     this._countries = values;
   }
 
+  Future<void> selectCountry(BuildContext context,
+      {String? initialShortName,
+      DialCodeSelectorTheme? selectorTheme,
+      Function(Country? selectedCountry)? onCountrySelected}) async {
+    Country? country = await showModalBottomSheet(
+        isScrollControlled: true,
+        backgroundColor: Colors.transparent,
+        barrierColor: Colors.black.withOpacity(0.75),
+        shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.vertical(
+                top: Radius.circular(selectorTheme?.sheetCornerRadius ?? 11))),
+        context: context,
+        builder: (context) {
+          return CountriesMainView(
+              dialCodeSelectorTheme: selectorTheme,
+              initialShortName: initialShortName);
+        });
+    if (country == null) return;
+    onCountrySelected?.call(Country.forCallBack(country));
+  }
+
   Country? getCountryByShortName({String? shortName}) =>
       Country.forCallBack(_getCountryByShortName(shortName: shortName));
 
@@ -32,7 +55,7 @@ class CountryViewModel {
     try {
       return _countries.singleWhere((element) =>
           element.dialCode?.replaceAll("+", "") ==
-          dialCode?.replaceAll("+", ""));
+          dialCode?.replaceAll("+", "").replaceAll(" ", ""));
     } catch (e) {
       return null;
     }
@@ -41,7 +64,8 @@ class CountryViewModel {
   Country? _getCountryByShortName({String? shortName}) {
     try {
       return _countries.singleWhere((element) =>
-          element.countryShortName?.toLowerCase() == shortName?.toLowerCase());
+          element.countryShortName?.toLowerCase() ==
+          shortName?.toLowerCase().replaceAll(" ", ""));
     } catch (e) {
       return null;
     }